requirement.txt
mecab-python3==0.996.1
